+++ samples/SlangpyTraining/NeuralModules.py
@@ -7,7 +7,7 @@
 # distribution of this software and related documentation without an express
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 
-from slangpy.backend import Device, CoopVecMatrixDesc, CoopVecMatrixLayout, ResourceUsage, DataType
+import slangpy as spy
 from slangpy.reflection import SlangType
 from slangpy.types import NDBuffer
 from typing import Any
@@ -18,7 +18,7 @@ import math
 # Root interface representing a slang type that implements the 
 # IModule interface for trainable CoopVec primitives
 class CoopVecModule:
-    def __init__(self, element_type: DataType, fan_in: int, fan_out: int):
+    def __init__(self, element_type: spy.DataType, fan_in: int, fan_out: int):
         super().__init__()
 
         self.element_type = element_type
@@ -131,7 +131,7 @@ class ModuleChain(CoopVecModule):
 # Frequency encoding that maps each input parameter into a series
 # of sines and cosines with increasing frequency
 class FrequencyEncoding(CoopVecModule):
-    def __init__(self, dtype: DataType, input_width: int, num_octaves: int):
+    def __init__(self, dtype: spy.DataType, input_width: int, num_octaves: int):
         self.num_octaves = num_octaves
         output_width = num_octaves * input_width * 2
         
@@ -161,7 +161,7 @@ class Activation:
     def __init__(self, act_name: str):
         self.act_name = act_name
 
-    def type_name(self, dtype: DataType, width: int) -> str:
+    def type_name(self, dtype: spy.DataType, width: int) -> str:
         return f"{self.act_name}<{dtype_name(dtype)}, {width}>"
     
     def get_this(self):
@@ -233,7 +233,7 @@ class TanhAct(Activation):
 # Stores the buffer and offset info that backs of multiple CoopVec matrices
 # and biases packed into a single structuredbuffer
 class CoopVecParams:
-    def __init__(self, device: Device, dtype: DataType, layout: CoopVecMatrixLayout, widths: tuple[int, ...]):
+    def __init__(self, device: spy.Device, dtype: spy.DataType, layout: spy.CoopVecMatrixLayout, widths: tuple[int, ...]):
         super().__init__()
 
         self.dtype = dtype
@@ -241,7 +241,7 @@ class CoopVecParams:
         self.elem_size = dtype_size(dtype)
 
         # First, compute offset and size info for the requested layout and layer widths
-        self.matrix_descs: list[CoopVecMatrixDesc] = []
+        self.matrix_descs: list[spy.CoopVecMatrixDesc] = []
         self.bias_offsets = []
         cur_offset = 0
         for fan_in, fan_out in zip(widths[:-1], widths[1:]):
@@ -255,7 +255,7 @@ class CoopVecParams:
             cur_offset += desc.rows * self.elem_size
 
         # Create the buffer
-        usage = ResourceUsage.shader_resource | ResourceUsage.unordered_access | ResourceUsage.shared
+        usage = spy.BufferUsage.shader_resource | spy.BufferUsage.unordered_access | spy.BufferUsage.shared
         self.elem_count = cur_offset // self.elem_size
         self.buffer = NDBuffer(device, dtype_name(dtype), self.elem_count, usage=usage)
 
@@ -266,11 +266,13 @@ class CoopVecParams:
         self.weights: list[np.ndarray] = []
         self.biases: list[np.ndarray] = []
         for b_offset, desc in zip(self.bias_offsets, self.matrix_descs):
-            np_bias = self.np_buffer[b_offset:b_offset + desc.rows*self.elem_size]
-            np_weight = self.np_buffer[desc.offset:desc.offset + desc.size]
+            b_index = b_offset // self.elem_size
+            m_index = desc.offset // self.elem_size
+            np_bias = self.np_buffer[b_index:b_index + desc.rows]
+            np_weight = self.np_buffer[m_index:m_index + desc.size // self.elem_size]
 
-            self.biases.append(np_bias.view(np_type))
-            self.weights.append(np_weight.view(np_type))
+            self.biases.append(np_bias)
+            self.weights.append(np_weight)
     
     # Copy weights from GPU to CPU
     def to_cpu(self):
@@ -283,7 +285,7 @@ class CoopVecParams:
     @property
     def weight_offsets(self):
         return [desc.offset for desc in self.matrix_descs]
-    
+
 
 # A trainable CoopVec MLP with configurable widths and activations
 # The hidden_act activation will be applied to every intermediate output,
@@ -293,8 +295,8 @@ class CoopVecParams:
 # with to_rowmajor and to_coopvec
 class TrainableMLP(CoopVecModule):
     def __init__(self,
-                 device: Device,
-                 dtype: DataType,
+                 device: spy.Device,
+                 dtype: spy.DataType,
                  num_hidden_layers: int,
                  input_width: int,
                  hidden_width: int,
@@ -319,25 +321,19 @@ class TrainableMLP(CoopVecModule):
         self.num_layers = len(self.widths) - 1
 
         # Create wo
-        self.params_rowmaj = CoopVecParams(device, dtype, CoopVecMatrixLayout.row_major, self.widths)
+        self.params_rowmaj = CoopVecParams(device, dtype, spy.CoopVecMatrixLayout.row_major, self.widths)
 
         # Initialize weights to random values and copy to row-major params
-        N = self.params_rowmaj.elem_count
-        np_params = np.zeros((N, ), dtype=dtype_to_numpy(dtype))
-        offsets = self.params_rowmaj.weight_offsets
-
         rng = np.random.default_rng(seed=12345)
-        for i, (start, end) in enumerate(zip(offsets, offsets[1:] + [np_params.nbytes])):
-            weight_slice = np_params[start//np_params.itemsize:end//np_params.itemsize]
-
+        for i, weight_slice in enumerate(self.params_rowmaj.weights):
             # Xavier uniform initialization
             std = math.sqrt(2.0 / (self.widths[i] + self.widths[i + 1]))
             a = math.sqrt(3.0) * std
             weight_slice[:] = rng.uniform(-a, a, (len(weight_slice), ))
-        self.params_rowmaj.buffer.copy_from_numpy(np_params)
+        self.params_rowmaj.to_gpu()
 
         # Create coopvec parameters and initialize
-        layout = CoopVecMatrixLayout.training_optimal
+        layout = spy.CoopVecMatrixLayout.training_optimal
         self.params_coopvec = CoopVecParams(device, dtype, layout, self.widths)
 
         param_buf = self.params_coopvec.buffer
@@ -410,55 +406,58 @@ class TrainableMLP(CoopVecModule):
 
     def serialize(self):
         result = {
-            'channels': self.widths,
-            'layers': {}
+            'layers': []
         }
 
         self.to_row_major()
         self.params_rowmaj.to_cpu()
 
         for i, (w, b) in enumerate(zip(self.params_rowmaj.weights, self.params_rowmaj.biases)):
-            result["layers"][f"layer_{i}_w"] = w.flatten().tolist()
-            result["layers"][f"layer_{i}_b"] = b.flatten().tolist()
+            result["layers"].append({
+                "num_inputs": self.widths[i],
+                "num_outputs": self.widths[i + 1],
+                "weights": w.flatten().tolist(),
+                "biases": b.flatten().tolist()
+            })
 
         return result
 
 
-def dtype_name(dtype: DataType):
-    if dtype == DataType.float16:
+def dtype_name(dtype: spy.DataType):
+    if dtype == spy.DataType.float16:
         return "half"
-    elif dtype == DataType.float32:
+    elif dtype == spy.DataType.float32:
         return "float"
     else:
         raise ValueError(f"Unsupported CoopVec datatype '{dtype}'")
     
 
-def dtype_to_numpy(dtype: DataType):
-    if dtype == DataType.float16:
+def dtype_to_numpy(dtype: spy.DataType):
+    if dtype == spy.DataType.float16:
         return np.float16
-    elif dtype == DataType.float32:
+    elif dtype == spy.DataType.float32:
         return np.float32
     else:
         raise ValueError(f"Unsupported CoopVec datatype '{dtype}'")
     
 
-def dtype_to_component_type(dtype: DataType):
-    if dtype == DataType.float16:
+def dtype_to_component_type(dtype: spy.DataType):
+    if dtype == spy.DataType.float16:
         return "CoopVecComponentType::Float16"
-    elif dtype == DataType.float32:
+    elif dtype == spy.DataType.float32:
         return "CoopVecComponentType::Float32"
     else:
         raise ValueError(f"Unsupported CoopVec datatype '{dtype}'")
 
 
-def dtype_size(dtype: DataType):
-    if dtype in (DataType.int8, DataType.uint8):
+def dtype_size(dtype: spy.DataType):
+    if dtype in (spy.DataType.int8, spy.DataType.uint8):
         return 1
-    elif dtype in (DataType.int16, DataType.uint16, DataType.float16):
+    elif dtype in (spy.DataType.int16, spy.DataType.uint16, spy.DataType.float16):
         return 2
-    elif dtype in (DataType.int32, DataType.uint32, DataType.float32):
+    elif dtype in (spy.DataType.int32, spy.DataType.uint32, spy.DataType.float32):
         return 4
-    elif dtype in (DataType.int64, DataType.uint64, DataType.float64):
+    elif dtype in (spy.DataType.int64, spy.DataType.uint64, spy.DataType.float64):
         return 8
     else:
         raise ValueError(f"Unsupported CoopVec datatype '{dtype}'")