+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # RTX Neural Shading Change Log
 
+## 1.2.0
+- Added Slangpy Inferencing sample.
+	- This sample demonstrates how to deploy a neural network prototyped with Python and SlangPy using C++ and Slang.
+- Updated samples to use NVRHI for Cooperative Vector format queries and matrix conversions.
+
 ## 1.1.0
 - Added DX12 cooperative vector support using Preview Agility SDK.
 - Moved matrix conversion to GPU.